+++ CHANGES.rst
@@ -25,6 +25,8 @@ Release history
 **Added**
 
 - Added official support for Python 3.9. (`#1660`_)
+- Added ``ChannelShape.from_space_and_channels`` to easily construct a ``ChannelShape``
+  from a spatial shape and number of channels. (`#1648`_)
 
 **Fixed**
 
@@ -44,6 +46,7 @@ Release history
 - Fixed a bug with a problematic cache index breaking decoder solvers. The solver now
   avoids using the cache, rather than crashing. (`#1649`_)
 
+.. _#1648: https://github.com/nengo/nengo/pull/1648
 .. _#1649: https://github.com/nengo/nengo/pull/1649
 .. _#1654: https://github.com/nengo/nengo/pull/1654
 .. _#1660: https://github.com/nengo/nengo/pull/1660